requirements.txt
psycopg2-binary==2.8.6
python-dotenv>=1.0.0
tkmacosx==1.0.5